Moved all tests from core/tests/test_api_views to curriculum_tracking/tests/test_api_views.py
